+++ example/example/apps/polls/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 
-from .models import Question
+from .models import Question, Choice
 
 admin.site.register(Question)
+admin.site.register(Choice)
