+++ example/example/controllers/home.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from fastapi import Depends
-from mountaineer import Metadata, RenderBase, ControllerBase
+from mountaineer import Metadata, RenderBase, ControllerBase, LayoutControllerBase
 from pydantic import BaseModel
 from starlette.requests import Request
 
@@ -16,28 +16,16 @@ class QuestionOutput(BaseModel):
         from_attributes = True
 
 
-class UserOutput(BaseModel):
-    username: str
-    email: str
-    first_name: str
-    last_name: str
-
-    class Config:
-        from_attributes = True
-
-
 class HomeRender(RenderBase):
     questions: list[QuestionOutput] = []
-    user: UserOutput | None = None
 
 
 class HomeController(ControllerBase):
     url = "/"
-    view_path = "src/home/page.tsx"
+    view_path = "src/pages/home/page.tsx"
 
     async def render(
             self,
-            user: UserOutput | None = Depends(AuthDependencies.get_user),
     ) -> HomeRender:
         from example.apps.polls.models import Question
 
@@ -46,6 +34,5 @@ class HomeController(ControllerBase):
                 QuestionOutput.from_orm(question)
                 async for question in Question.objects.all()
             ],
-            user=user,
             metadata=Metadata(title="Home"),
         )
