+++ example/example/views/tailwind.config.js
@@ -1,9 +1,10 @@
+import daisyui from "daisyui"
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: ["./src/**/*.{html,js,ts,jsx,tsx}"],
-  theme: {
-    extend: {},
+  daisyui: {
+    themes: ["night"],
   },
-  plugins: [],
+  plugins: [daisyui],
 };
